# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkApplyRequest": {
        "properties": {
          "confirm": {
            "default": false,
            "description": "must be true",
            "title": "Confirm",
            "type": "boolean"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "fields to write; omit to use each item's default selection, which is empty for anything needing review",
            "title": "Fields"
          },
          "job_id": {
            "description": "from a /bulk/diff run",
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "BulkApplyRequest",
        "type": "object"
      },
      "BulkDiffRequest": {
        "properties": {
          "min_confidence": {
            "default": 0.0,
            "description": "skip items whose best candidate scores below this",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "type": "number"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/FieldPolicyOverride"
            },
            "title": "Overrides",
            "type": "array"
          },
          "selection": {
            "$ref": "#/components/schemas/BulkSelection",
            "default": {
              "kind": "artist",
              "limit": 50,
              "missing_metadata": false
            }
          },
          "tag_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TagPolicyRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "BulkDiffRequest",
        "type": "object"
      },
...
      "BulkSelection": {
        "description": "Which items, and nothing else. The fields to write are chosen after review.",
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "explicit item ids, if known",
            "title": "Ids"
          },
          "kind": {
            "default": "artist",
            "enum": [
              "artist",
              "album",
              "song"
            ],
            "title": "Kind",
            "type": "string"
          },
          "limit": {
            "default": 50,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "missing_metadata": {
            "default": false,
            "description": "only items lacking genres, provider ids or an overview",
            "title": "Missing Metadata",
            "type": "boolean"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "search": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search"
          }
        },
        "title": "BulkSelection",
        "type": "object"
      },
...
      "FieldPolicyOverride": {
        "description": "Per-field mode for one request, so a single edit need not change configuration.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "mode": {
            "enum": [
              "keep_existing",
              "fill_if_empty",
              "replace",
              "merge"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "field",
          "mode"
        ],
        "title": "FieldPolicyOverride",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TagPolicyRequest": {
        "properties": {
          "extra_blacklist": {
            "items": {
              "type": "string"
            },
            "title": "Extra Blacklist",
            "type": "array"
          },
          "genre_limit": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Genre Limit",
            "type": "integer"
          },
          "min_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Min Count",
            "type": "integer"
          },
          "style_limit": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Style Limit",
            "type": "integer"
          }
        },
        "title": "TagPolicyRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Browse a Jellyfin music library, read artist/album/track metadata from the\nLast.fm API, keep a persistent incremental local archive of every Last.fm\nresponse, review a field-level diff, and apply the accepted changes back to\nJellyfin.\n",
    "title": "metaedit",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bulk/apply": {
      "post": {
        "description": "Apply a reviewed job, streaming progress and isolating failures per item.\n\n``confirm`` must be true. The job must exist, which means a diff was reviewed in\nthis process: a bulk write is never the first request of a session.",
        "operationId": "bulk_apply_api_bulk_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Apply",
        "tags": [
          "bulk"
        ]
      }
    },
    "/api/bulk/diff": {
      "post": {
        "description": "Diff every item in the selection, streaming one event per item.\n\n**Writes nothing.** The job id in the summary event is what a later ``/bulk/apply``\nmust present, so an apply can never be the first thing that happens.",
        "operationId": "bulk_diff_api_bulk_diff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDiffRequest",
                "default": {
                  "min_confidence": 0.0,
                  "overrides": [],
                  "selection": {
                    "kind": "artist",
                    "limit": 50,
                    "missing_metadata": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Diff",
        "tags": [
          "bulk"
        ]
      }
    },
...
    "/api/bulk/{batch_id}/revert": {
      "post": {
        "description": "Undo every item written by a batch, oldest first, isolating failures.\n\nThe batch id lives on the snapshots, so this works after a restart even though the\ndiff job that produced it does not.",
        "operationId": "bulk_revert_api_bulk__batch_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "default": false,
              "title": "Confirm",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Revert",
        "tags": [
          "bulk"
        ]
      }
    },
...
  }
}